requirements-test.txt
# Minimal dependencies for testing in CI/CD
# Core Testing
pytest>=7.4.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Web Frameworks (lightweight, no heavy ML deps)
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0

# Database
pymongo>=4.6.0
redis>=5.0.0

# API Client
openai>=1.0.0
httpx>=0.27.0
requests>=2.31.0

# Data Processing (minimal)
numpy<2.0.0,>=1.20.0
pandas>=2.0.0

# Utilities
pyyaml>=6.0.0
pydantic<3.0.0,>=2.0.0
tqdm>=4.65.0
